requirements.txt
torch>=2.0
torchvision>=0.15
diffusers>=0.25
numpy
scipy
Pillow
omegaconf
tqdm
wandb
matplotlib
pyserial
pytest
